agentcore/requirements.txt
strands-agents>=0.1.0
boto3>=1.34.0
botocore>=1.34.0
pydantic>=2.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
mangum>=0.17.0
